Add support extra Schemas
