+++ openapi-annotations/src/main/java/io/micronaut/openapi/annotation/OpenAPIExtraSchemas.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2017-2023 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micronaut.openapi.annotation;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.RetentionPolicy.SOURCE;
+
+/**
+ * Allows {@link OpenAPIExtraSchema} to be repeatable.
+ *
+ * @since 6.12.0
+ */
+@Documented
+@Retention(SOURCE)
+@Target({ElementType.PACKAGE, ElementType.TYPE})
+public @interface OpenAPIExtraSchemas {
+
+    /**
+     * @return A group of {@link OpenAPIExtraSchema}
+     */
+    OpenAPIExtraSchema[] value() default {};
+}
