+++ openapi/src/main/java/io/micronaut/openapi/visitor/OpenApiConfigProperty.java
@@ -48,6 +48,10 @@ public interface OpenApiConfigProperty {
      */
     String MICRONAUT_OPENAPI_CONFIG_FILE = "micronaut.openapi.config.file";
     /**
+     * System property that enables extra schema processing.
+     */
+    String MICRONAUT_OPENAPI_EXTRA_SCHEMA_ENABLED = "micronaut.openapi.extra.schema.enabled";
+    /**
      * Prefix for expandable properties.
      */
     String MICRONAUT_OPENAPI_EXPAND_PREFIX = "micronaut.openapi.expand.";
@@ -250,30 +254,31 @@ public interface OpenApiConfigProperty {
      * All supported annotation processor properties.
      */
     Set<String> ALL = Set.of(
-        MICRONAUT_OPENAPI_ENABLED,
-        MICRONAUT_OPENAPI_CONTEXT_SERVER_PATH,
-        MICRONAUT_OPENAPI_PROPERTY_NAMING_STRATEGY,
-        MICRONAUT_OPENAPI_VIEWS_SPEC,
-        MICRONAUT_OPENAPI_FILENAME,
-        MICRONAUT_OPENAPI_JSON_FORMAT,
-        MICRONAUT_OPENAPI_ENVIRONMENTS,
-        MICRONAUT_ENVIRONMENT_ENABLED,
-        MICRONAUT_OPENAPI_FIELD_VISIBILITY_LEVEL,
-        MICRONAUT_CONFIG_FILE_LOCATIONS,
-        MICRONAUT_OPENAPI_TARGET_FILE,
-        MICRONAUT_OPENAPI_VIEWS_DEST_DIR,
-        MICRONAUT_OPENAPI_ADDITIONAL_FILES,
-        MICRONAUT_OPENAPI_CONFIG_FILE,
-        MICRONAUT_OPENAPI_SECURITY_ENABLED,
-        MICRONAUT_OPENAPI_VERSIONING_ENABLED,
-        MICRONAUT_OPENAPI_JSON_VIEW_DEFAULT_INCLUSION,
-        MICRONAUT_OPENAPI_PROJECT_DIR,
-        MICRONAUT_OPENAPI_ADOC_ENABLED,
-        MICRONAUT_OPENAPI_ADOC_TEMPLATES_DIR_PATH,
-        MICRONAUT_OPENAPI_ADOC_TEMPLATE_FILENAME,
-        MICRONAUT_OPENAPI_ADOC_OUTPUT_DIR_PATH,
-        MICRONAUT_OPENAPI_ADOC_OUTPUT_FILENAME,
-        MICRONAUT_OPENAPI_ADOC_OPENAPI_PATH,
-        MICRONAUT_OPENAPI_SWAGGER_FILE_GENERATION_ENABLED
+            MICRONAUT_OPENAPI_ENABLED,
+            MICRONAUT_OPENAPI_CONTEXT_SERVER_PATH,
+            MICRONAUT_OPENAPI_PROPERTY_NAMING_STRATEGY,
+            MICRONAUT_OPENAPI_VIEWS_SPEC,
+            MICRONAUT_OPENAPI_FILENAME,
+            MICRONAUT_OPENAPI_JSON_FORMAT,
+            MICRONAUT_OPENAPI_ENVIRONMENTS,
+            MICRONAUT_ENVIRONMENT_ENABLED,
+            MICRONAUT_OPENAPI_FIELD_VISIBILITY_LEVEL,
+            MICRONAUT_CONFIG_FILE_LOCATIONS,
+            MICRONAUT_OPENAPI_TARGET_FILE,
+            MICRONAUT_OPENAPI_VIEWS_DEST_DIR,
+            MICRONAUT_OPENAPI_ADDITIONAL_FILES,
+            MICRONAUT_OPENAPI_CONFIG_FILE,
+            MICRONAUT_OPENAPI_SECURITY_ENABLED,
+            MICRONAUT_OPENAPI_VERSIONING_ENABLED,
+            MICRONAUT_OPENAPI_JSON_VIEW_DEFAULT_INCLUSION,
+            MICRONAUT_OPENAPI_PROJECT_DIR,
+            MICRONAUT_OPENAPI_ADOC_ENABLED,
+            MICRONAUT_OPENAPI_ADOC_TEMPLATES_DIR_PATH,
+            MICRONAUT_OPENAPI_ADOC_TEMPLATE_FILENAME,
+            MICRONAUT_OPENAPI_ADOC_OUTPUT_DIR_PATH,
+            MICRONAUT_OPENAPI_ADOC_OUTPUT_FILENAME,
+            MICRONAUT_OPENAPI_ADOC_OPENAPI_PATH,
+            MICRONAUT_OPENAPI_SWAGGER_FILE_GENERATION_ENABLED,
+            MICRONAUT_OPENAPI_EXTRA_SCHEMA_ENABLED
     );
 }
