+++ openapi/src/main/resources/META-INF/services/io.micronaut.inject.visitor.TypeElementVisitor
@@ -4,3 +4,4 @@ io.micronaut.openapi.visitor.OpenApiEndpointVisitor
 io.micronaut.openapi.visitor.OpenApiIncludeVisitor
 io.micronaut.openapi.visitor.OpenApiJacksonVisitor
 io.micronaut.openapi.visitor.OpenApiGroupInfoVisitor
+io.micronaut.openapi.visitor.OpenApiExtraSchemaVisitor
