+++ openapi/src/test/groovy/io/micronaut/openapi/extra/ExtraModel1.java
@@ -1,0 +1,12 @@
+package io.micronaut.openapi.extra;
+
+public class ExtraModel1 {
+
+    public String prop1;
+    public InternalClass ddd;
+
+    public static class InternalClass {
+
+        public String prop2;
+    }
+}
