+++ openapi/src/test/groovy/io/micronaut/openapi/extra/ExtraModel2.java
@@ -1,0 +1,6 @@
+package io.micronaut.openapi.extra;
+
+public class ExtraModel2 {
+
+    public int field1;
+}
