+++ openapi/src/test/groovy/io/micronaut/openapi/extra/internal/ExtraModelInternalPackage.java
@@ -1,0 +1,6 @@
+package io.micronaut.openapi.extra.internal;
+
+public class ExtraModelInternalPackage {
+
+    public String prop1;
+}
